chore - refactored new utils into seperate files plus other minor change from code review
