+++ bc_obps/service/utils/constants.py
@@ -1,0 +1,3 @@
+# May 31st. Month and Day appended to reporting_year_id(calendar year as an int) in get_report_valid_year_from_version_id
+# to form a date string used to find the valid configuration for a report version
+REPORT_VERSION_DATE_SUFFIX = "-05-31"
